- cree e implemente las rutas   - instale react-router-dom   - elimine comentarios no necesarios - Las rutas creadas las llame Home y ComponentsList.
