+++ package.json
@@ -12,6 +12,7 @@
     "prop-types": "^15.8.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "reactstrap": "^9.0.1",
     "web-vitals": "^2.1.2"
