+++ public/index.html
@@ -28,24 +28,12 @@
   </head>
   <body class="d-flex flex-column h-100">
     <noscript>You need to enable JavaScript to run this app.</noscript>
- 
     <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
-
+    <script type="text/babel" src="../src/index.js"></script>
     <footer class="footer mt-auto py-3 bg-light">
       <div class="container">
         <span class="text-muted">Hecho por Sol</span>
       </div>
     </footer>
-      
   </body>
 </html>
